+++ app/src/hooks/useFetchTidal.ts
@@ -1,4 +1,5 @@
 import { useState } from "react";
+import { TIDAL_API_URL, TIDARR_PROXY_URL } from "src/contants";
 import { useConfigProvider } from "src/provider/ConfigProvider";
 import { ConfigTiddleType } from "src/types";
 
@@ -8,13 +9,12 @@ async function fetchTidal<T>(
   url: string,
   options: RequestInit = {},
   tiddlConfig?: ConfigTiddleType,
+  useProxy?: boolean,
 ): Promise<T | undefined> {
   const countryCode = tiddlConfig?.auth.country_code || "EN";
   const TOKEN = tiddlConfig?.auth.token;
-  const apiUrl =
-    import.meta.env.MODE === "development"
-      ? import.meta.env.VITE_TIDARR_PROXY_URL
-      : "/proxy";
+
+  const apiUrl = useProxy ? TIDARR_PROXY_URL : TIDAL_API_URL;
 
   options.headers = new Headers({
     ...options?.headers,
@@ -41,12 +41,17 @@ async function fetchTidal<T>(
 }
 
 export function useFetchTidal() {
-  const { tiddlConfig } = useConfigProvider();
+  const { tiddlConfig, config } = useConfigProvider();
   const [loading, setLoading] = useState<boolean>(false);
 
   async function fetcher<T>(url: string, options?: RequestInit) {
     setLoading(true);
-    const data = await fetchTidal<T>(url, options, tiddlConfig);
+    const data = await fetchTidal<T>(
+      url,
+      options,
+      tiddlConfig,
+      config?.ENABLE_TIDAL_PROXY === "true",
+    );
     setLoading(false);
     return data;
   }